requirements.txt
# Odoo 17 Community Edition
odoo==17.0

# Core Dependencies
Babel==2.9.1
chardet==4.0.0
cryptography==41.0.0
decorator==4.4.2
docutils==0.17
ebaysdk==2.1.5
freezegun==1.1.0
gevent==22.10.2
greenlet==2.0.1
idna==2.10
Jinja2==3.0.3
libsass==0.20.1
lxml==4.9.2
MarkupSafe==2.1.0
num2words==0.5.10
ofxparse==0.21
passlib==1.7.4
Pillow==9.4.0
polib==1.1.1
psutil==5.9.4
psycopg2-binary==2.9.5
pydot==1.4.2
pyopenssl==23.0.0
PyPDF2==1.26.0
pyserial==3.5
python-dateutil==2.8.1
python-ldap==3.4.0
python-stdnum==1.17
pytz==2022.7.1
pyusb==1.2.1
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.2
urllib3==1.26.14
vobject==0.9.6.1
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==3.0.8
xlwt==1.3.0
zeep==4.1.0

# Additional useful packages
python-dotenv==0.19.2
